fixes after review on GitHub
