+++ sampo/schemas/time_estimator.py
@@ -2,7 +2,6 @@ from abc import ABC, abstractmethod
 from enum import Enum
 from operator import attrgetter
 from random import Random
-from typing import Optional
 
 import numpy.random
 import math
@@ -115,24 +114,6 @@ class DefaultWorkEstimator(WorkTimeEstimator):
             self._productivity[name]['__ALL__'] = productivity
             return
         self._productivity[name][contractor] = productivity
-
-
-class PSPlibWorkTimeEstimator(WorkTimeEstimator):
-    def __init__(self, times: list[Time]):
-        self.times = times
-
-    def find_work_resources(self, work_name: str, work_volume: float, resource_name: list[str] | None = None) \
-            -> dict[str, int]:
-        ...
-
-    def set_estimation_mode(self, use_idle: bool = True, mode: WorkEstimationMode = WorkEstimationMode.Realistic):
-        ...
-
-    def set_productivity_mode(self, mode: WorkerProductivityMode = WorkerProductivityMode.Static):
-        ...
-
-    def estimate_time(self, work_unit: WorkUnit, worker_list: list[Worker]) -> Time:
-        return Time(work_unit.time_exec)
 
 
 def communication_coefficient(groups_count: int, max_groups: int) -> float:
